renamed FooterWrapper file to match export name, updated imports
