+++ src/components/Footer/Footer.tsx
@@ -6,7 +6,7 @@
  *
  */
 
-import React, { ReactNode } from "react";
+import { ReactNode } from "react";
 import styled from "@emotion/styled";
 import { colors, spacing, stackOrder } from "@ndla/core";
 import { MAX_PAGE_WIDTH } from "../../constants";
