+++ src/containers/Login/Login.tsx
@@ -15,7 +15,7 @@ import { OneColumn } from "@ndla/ui";
 import LoginFailure from "./LoginFailure";
 import LoginProviders from "./LoginProviders";
 import LoginSuccess from "./LoginSuccess";
-import Footer from "../App/components/Footer";
+import Footer from "../App/components/FooterWrapper";
 import { useSession } from "../Session/SessionProvider";
 
 const StyledOneColumn = styled(OneColumn)`
