+++ src/containers/NdlaFilm/NdlaFilm.tsx
@@ -9,7 +9,7 @@
 import { Route, Routes } from "react-router-dom";
 import { OneColumn } from "@ndla/ui";
 import NdlaFilmEditor from "./NdlaFilmEditor";
-import Footer from "../App/components/Footer";
+import Footer from "../App/components/FooterWrapper";
 
 const NdlaFilm = () => {
   return (
